+++ docs/changelog.md
@@ -5,6 +5,7 @@
 - DSL identifiers (if present) will now be loaded when extending a JSON workspace (see https://github.com/structurizr/dsl/discussions/328).
 - Adds a `context` variable to inline/external scripts (see https://github.com/structurizr/dsl/issues/332).
 - Fixes https://github.com/structurizr/dsl/issues/324 (Groups with no curly braces breaks diagrams).
+- Adds a way to set the character encoding used by the DSL parser (see https://github.com/structurizr/dsl/issues/338).
 
 ## 1.32.0 (28th July 2023)